requirements.txt
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28
typing_extensions==4.10.0
